Fix cuda moving

+++ train.py
@@ -139,8 +139,8 @@ for epoch_num in range(opt.epoch, opt.n_epochs):
             #to_cuda(real_batch_A)
             #to_cuda(real_batch_B)
 
-            real_batch_A['source_tokens']['onehots'].cuda()
-            real_batch_A['source_tokens']['ids'].cuda()
+            real_batch_A['source_tokens']['onehots'].float().cuda()
+            real_batch_A['source_tokens']['ids'].float().cuda()
 
             print(real_batch_A['source_tokens']['onehots'])
